requirements.txt
git+https://github.com/khaitruong922/discord-bot@rewrite
discord
requests
python-dotenv